+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+## [v3.2.0] - 2026-07-07
+
+Added:
+
+- Refactored popup as Vue 3 app
+- Refactored main content script that manages frame creation and navigation to be a Vue 3 app
+- Refactored frame serialisation to use indexed parameters
+
 ## [v3.1.0] - 2026-07-05
 
 Added: